переименовал функцию создания данных, getAvatarNumber и getPhotos перенес в data.js
