+++ js/util.js
@@ -1,6 +1,3 @@
-const START_RANGE = 1;
-const END_RANGE = 10;
-
 function getRandomPositiveInteger (a, b) {
   const lower = Math.ceil(Math.min(Math.abs(a), Math.abs(b)));
   const upper = Math.floor(Math.max(Math.abs(a), Math.abs(b)));
@@ -31,15 +28,4 @@ function getRandomStringFromArray(elements) {
   return mixElements(elements).slice(0, getRandomPositiveInteger(1, elements.length - 1)).join(', ');
 }
 
-function getAvatarNumber () {
-  const number = getRandomPositiveInteger(START_RANGE, END_RANGE);
-  return number < 10 ? `0${number}` : number;
-}
-
-function getPhotos (elements) {
-  const elementItem = () => getRandomArrayElement(elements);
-  const photos = Array.from({length: getRandomPositiveInteger (3, 10)}, elementItem);
-  return photos;
-}
-
-export {getRandomPositiveInteger, getRandomPositiveFloat, getRandomArrayElement, getRandomStringFromArray, getAvatarNumber, getPhotos};
+export {getRandomPositiveInteger, getRandomPositiveFloat, getRandomArrayElement, getRandomStringFromArray};
